+++ src/test/kotlin/com/jetbrains/snakecharm/SnakemakeTestCase.kt
@@ -1,8 +1,10 @@
 package com.jetbrains.snakecharm
 
 import com.intellij.application.options.CodeStyle
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.roots.impl.FilePropertyPusher
+import com.intellij.openapi.vfs.newvfs.impl.VfsRootAccess
 import com.intellij.testFramework.TestApplicationManager
 import com.intellij.testFramework.UsefulTestCase
 import com.intellij.testFramework.fixtures.CodeInsightTestFixture
@@ -43,6 +45,9 @@ abstract class SnakemakeTestCase : UsefulTestCase() {
         super.setUp()
         TestApplicationManager.getInstance()
         val factory = IdeaTestFixtureFactory.getFixtureFactory()
+
+        allowPythonRootsAccess(testRootDisposable)
+
         val fixtureBuilder = factory.createLightFixtureBuilder(
             projectDescriptor, getTestName(false)
         )
@@ -51,6 +56,8 @@ abstract class SnakemakeTestCase : UsefulTestCase() {
                 createTempDirFixture()
         )
         fixture!!.testDataPath = SnakemakeTestUtil.getTestDataPath().toString()
+        VfsRootAccess.allowRootAccess(testRootDisposable, fixture!!.testDataPath)
+
         // TODO() replace with: runInEdtAndWait
         if (SwingUtilities.isEventDispatchThread()) {
             fixture!!.setUp()
@@ -127,4 +134,18 @@ abstract class SnakemakeTestCase : UsefulTestCase() {
 
     protected open fun getIndentOptions() = getCommonCodeStyleSettings().indentOptions
 
+    companion object {
+        fun allowPythonRootsAccess(disposable: Disposable) {
+            // It is ok to access these pythons
+            //  Fix for: ERROR: File accessed outside allowed roots: file:///usr/local/bin/python3;
+
+            VfsRootAccess.allowRootAccess(
+                disposable,
+                "/usr/local/bin/python3.10",
+                "/usr/local/bin/python3",
+                "/usr/bin/python3",
+                "/usr/bin/python3.10"
+            )
+        }
+    }
 }