+++ src/test/kotlin/features/glue/FilesSteps.kt
@@ -70,10 +70,13 @@ class FilesSteps {
     }
 
     private fun createAndAddFile(name: String, text: String) {
+        // XXX: Cannot simplify to:
+        //      SnakemakeWorld.fixture().configureByText(name, StringUtil.convertLineSeparators(text))
+        // because it doesn't support relative paths, e.g. like "rules/foo.smk"
         ApplicationManager.getApplication().invokeAndWait({
             ApplicationManager.getApplication().runWriteAction {
                 val file = SnakemakeWorld.fixture().addFileToProject(
-                        name, StringUtil.convertLineSeparators(text)
+                    name, StringUtil.convertLineSeparators(text)
                 )
                 SnakemakeWorld.fixture().configureFromExistingVirtualFile(file.virtualFile)
             }